simulator/requirements.txt
prometheus_client==0.22.1
psycopg2-binary==2.9.10
tabulate==0.9.0
